+++ lib/avtp_pipeline/avtp_pipeline.mk
@@ -1,6 +1,7 @@
 AVB_FEATURE_ENDPOINT ?= 1
 IGB_LAUNCHTIME_ENABLED ?= 0
 AVB_FEATURE_GSTREAMER ?= 1
+PLATFORM_TOOLCHAIN ?= x86_i210_linux
 
 .PHONY: all clean
 
@@ -18,7 +19,7 @@ clean:
 build/Makefile:
 	mkdir -p build && \
 	cd build && \
-	cmake -DCMAKE_TOOLCHAIN_FILE=../platform/Linux/x86_i210_linux.cmake \
+	cmake -DCMAKE_TOOLCHAIN_FILE=../platform/Linux/$(PLATFORM_TOOLCHAIN).cmake \
 	      -DAVB_FEATURE_ENDPOINT=$(AVB_FEATURE_ENDPOINT) \
 	      -DIGB_LAUNCHTIME_ENABLED=$(IGB_LAUNCHTIME_ENABLED) \
 	      -DAVB_FEATURE_GSTREAMER=$(AVB_FEATURE_GSTREAMER) \
