+++ lib/avtp_pipeline/platform/Linux/openavb_osal.c
@@ -29,7 +29,6 @@ https://github.com/benhoyt/inih/commit/74d2ca064fb293bc60a77b0bd068075b293cf175.
 *************************************************************************************************************/
 
 #include "openavb_platform.h"
-#include "openavb_ether_hal.h"
 #include "openavb_osal.h"
 #include "openavb_qmgr.h"
