+++ lib/avtp_pipeline/platform/Linux/openavb_time_osal.c
@@ -1,16 +1,16 @@
 /*************************************************************************************************************
 Copyright (c) 2012-2015, Symphony Teleca Corporation, a Harman International Industries, Incorporated company
 All rights reserved.
- 
+
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are met:
- 
+
 1. Redistributions of source code must retain the above copyright notice, this
    list of conditions and the following disclaimer.
 2. Redistributions in binary form must reproduce the above copyright notice,
    this list of conditions and the following disclaimer in the documentation
    and/or other materials provided with the distribution.
- 
+
 THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS LISTED "AS IS" AND
 ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
 WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -21,10 +21,10 @@ LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
 ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- 
-Attributions: The inih library portion of the source code is licensed from 
-Brush Technology and Ben Hoyt - Copyright (c) 2009, Brush Technology and Copyright (c) 2009, Ben Hoyt. 
-Complete license and copyright information can be found at 
+
+Attributions: The inih library portion of the source code is licensed from
+Brush Technology and Ben Hoyt - Copyright (c) 2009, Brush Technology and Copyright (c) 2009, Ben Hoyt.
+Complete license and copyright information can be found at
 https://github.com/benhoyt/inih/commit/74d2ca064fb293bc60a77b0bd068075b293cf175.
 *************************************************************************************************************/
 
@@ -33,12 +33,10 @@ https://github.com/benhoyt/inih/commit/74d2ca064fb293bc60a77b0bd068075b293cf175.
 #include <sys/mman.h>
 #include <sys/ioctl.h>
 
-#include "igb.h"
-#include "avb.h"
+#include "avb_gptp.h"
 
 #include "openavb_platform.h"
 #include "openavb_time_osal.h"
-#include "openavb_time_hal.h"
 #include "openavb_trace.h"
 
 #define	AVB_LOG_COMPONENT	"osalTime"
@@ -52,26 +50,20 @@ static pthread_mutex_t gOSALTimeInitMutex = PTHREAD_MUTEX_INITIALIZER;
 #define UNLOCK()	pthread_mutex_unlock(&gOSALTimeInitMutex)
 
 static bool bInitialized = FALSE;
-static int gIgbShmFd = -1;
-static char *gIgbMmap = NULL;
+static int gPtpShmFd = -1;
+static char *gPtpMmap = NULL;
 gPtpTimeData gPtpTD;
 
 static bool x_timeInit(void) {
 	AVB_TRACE_ENTRY(AVB_TRACE_TIME);
 
-	if (!halTimeInitialize()) {
-		AVB_LOG_ERROR("HAL Time Init failed");
-		AVB_TRACE_EXIT(AVB_TRACE_TIME);
-		return FALSE;
-	}
-
-	if (gptpinit(&gIgbShmFd, &gIgbMmap) < 0) {
+	if (gptpinit(&gPtpShmFd, &gPtpMmap) < 0) {
 		AVB_LOG_ERROR("GPTP init failed");
 		AVB_TRACE_EXIT(AVB_TRACE_TIME);
 		return FALSE;
 	}
 
-	if (gptpgetdata(gIgbMmap, &gPtpTD) < 0) {
+	if (gptpgetdata(gPtpMmap, &gPtpTD) < 0) {
 		AVB_LOG_ERROR("GPTP data fetch failed");
 		AVB_TRACE_EXIT(AVB_TRACE_TIME);
 		return FALSE;
@@ -88,7 +80,7 @@ static bool x_timeInit(void) {
 static bool x_getPTPTime(U64 *timeNsec) {
 	AVB_TRACE_ENTRY(AVB_TRACE_TIME);
 
-	if (gptpgetdata(gIgbMmap, &gPtpTD) < 0) {
+	if (gptpgetdata(gPtpMmap, &gPtpTD) < 0) {
 		AVB_LOG_ERROR("GPTP data fetch failed");
 		AVB_TRACE_EXIT(AVB_TRACE_TIME);
 		return FALSE;
@@ -130,9 +122,7 @@ bool osalAVBTimeInit(void) {
 bool osalAVBTimeClose(void) {
 	AVB_TRACE_ENTRY(AVB_TRACE_TIME);
 
-	gptpdeinit(&gIgbShmFd, &gIgbMmap);
-
-	halTimeFinalize();
+	gptpdeinit(&gPtpShmFd, &gPtpMmap);
 
 	AVB_TRACE_EXIT(AVB_TRACE_TIME);
 	return TRUE;
@@ -203,4 +193,3 @@ bool osalClockGettime64(openavb_clockId_t openavbClockId, U64 *timeNsec) {
 	AVB_TRACE_EXIT(AVB_TRACE_TIME);
 	return FALSE;
 }
-
