+++ lib/avtp_pipeline/qmgr/openavb_qmgr.c
@@ -1,16 +1,16 @@
 /*************************************************************************************************************
 Copyright (c) 2012-2015, Symphony Teleca Corporation, a Harman International Industries, Incorporated company
 All rights reserved.
- 
+
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are met:
- 
+
 1. Redistributions of source code must retain the above copyright notice, this
    list of conditions and the following disclaimer.
 2. Redistributions in binary form must reproduce the above copyright notice,
    this list of conditions and the following disclaimer in the documentation
    and/or other materials provided with the distribution.
- 
+
 THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS LISTED "AS IS" AND
 ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
 WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -21,10 +21,10 @@ LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
 ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- 
-Attributions: The inih library portion of the source code is licensed from 
-Brush Technology and Ben Hoyt - Copyright (c) 2009, Brush Technology and Copyright (c) 2009, Ben Hoyt. 
-Complete license and copyright information can be found at 
+
+Attributions: The inih library portion of the source code is licensed from
+Brush Technology and Ben Hoyt - Copyright (c) 2009, Brush Technology and Copyright (c) 2009, Ben Hoyt.
+Complete license and copyright information can be found at
 https://github.com/benhoyt/inih/commit/74d2ca064fb293bc60a77b0bd068075b293cf175.
 *************************************************************************************************************/
 
@@ -33,7 +33,7 @@ https://github.com/benhoyt/inih/commit/74d2ca064fb293bc60a77b0bd068075b293cf175.
  */
 
 #include <openavb_types.h>
-#include "openavb_ether_hal.h"
+
 #define AVB_LOG_COMPONENT "QMGR"
 //#define AVB_LOG_LEVEL AVB_LOG_LEVEL_DEBUG
 #include "openavb_log.h"
@@ -41,7 +41,10 @@ https://github.com/benhoyt/inih/commit/74d2ca064fb293bc60a77b0bd068075b293cf175.
 
 #include "openavb_qmgr.h"
 #include "avb_sched.h"
-#include "openavb_ether_hal.h"
+
+#if (AVB_FEATURE_IGB)
+#include "openavb_igb.h"
+#endif
 
 #define AVB_DEFAULT_QDISC_MODE AVB_SHAPER_HWQ_PER_CLASS
 
@@ -49,7 +52,9 @@ https://github.com/benhoyt/inih/commit/74d2ca064fb293bc60a77b0bd068075b293cf175.
 
 // Qdisc configuration
 typedef struct {
+#if (AVB_FEATURE_IGB)
 	device_t *igb_dev;
+#endif
 	int mode;
 	int ifindex;
 	char ifname[IFNAMSIZ];
@@ -103,10 +108,11 @@ static bool setupHWQueue(int nClass, unsigned classBytesPerSec)
 		class_a_bytes_per_sec =  qmgr_classes[SR_CLASS_A].classBytesPerSec;
 		class_b_bytes_per_sec = classBytesPerSec;
 	}
-
+#if (AVB_FEATURE_IGB)
 	err = igb_set_class_bandwidth2(qdisc_data.igb_dev, class_a_bytes_per_sec, class_b_bytes_per_sec);
 	if (err)
 		AVB_LOGF_ERROR("Adding stream; igb_set_class_bandwidth failed: %s", strerror(err));
+#endif
 
 	AVB_TRACE_EXIT(AVB_TRACE_QUEUE_MANAGER);
 	return !err;
@@ -246,12 +252,14 @@ bool openavbQmgrInitialize(int mode, int ifindex, const char* ifname, unsigned m
 	AVB_LOGF_DEBUG("Initializing QMgr; mode=%d, idx=%d, mtu=%u, link_kbit=%u, nsr_kbit=%u",
 				   qdisc_data.mode, ifindex, mtu, link_kbit, nsr_kbit);
 
+#if (AVB_FEATURE_IGB)
 	if ( qdisc_data.mode != AVB_SHAPER_DISABLED
 	     && (qdisc_data.igb_dev = igbAcquireDevice()) == 0)
 	{
 		AVB_LOG_ERROR("Initializing QMgr; unable to acquire igb device");
 	}
 	else
+#endif
 	{
 		// Initialize data for classes and streams
 		memset(qmgr_classes, 0, sizeof(qmgr_classes));
@@ -294,8 +302,11 @@ void openavbQmgrFinalize(void)
 				}
 			}
 		}
+
+#if (AVB_FEATURE_IGB)
 		igbReleaseDevice(qdisc_data.igb_dev);
 		qdisc_data.igb_dev = NULL;
+#endif
 	}
 
 	UNLOCK();
